Move back to package by layer for now, don't mix both systems.
