+++ app/src/main/java/be/ugent/zeus/hydra/models/library/OpeningHours.java
@@ -1,4 +1,4 @@
-package be.ugent.zeus.hydra.library.details;
+package be.ugent.zeus.hydra.models.library;
 
 import be.ugent.zeus.hydra.models.converters.DateThreeTenAdapter;
 import be.ugent.zeus.hydra.models.converters.ZonedThreeTenAdapter;
