+++ app/src/main/res/layout/activity_library_details.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     tools:targetApi="lollipop"
-    tools:context=".library.details.DetailActivity">
+    tools:context=".activities.LibraryDetailActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/app_bar_layout"
